première version du main info box
Update .gitignore to exclude .flatpak-builder directory

The .gitignore file has been updated to include a new entry:
/wpfetch_compagnon_app.Desktop/.flatpak-builder. This change
ensures that the specified directory and its contents will be
ignored by Git, preventing them from being tracked or included
in version control. This is useful for excluding build artifacts
or other files that are not necessary to track in the repository.